# src/server/api/swagger.json
{
  "swagger": "2.0",
  "produces": [
    "application/json"
  ],
  "info": {
    "title": "conversationdiscovery",
    "description": "An API",
    "version": "1.0"
  },
  "basePath": "/conversationdiscovery",
  "schemes": [
    "http",
    "https"
  ],
  "parameters": {
    "clientId": {
      "name": "clientId",
      "in": "path",
      "description": "A client id, e.g. \"247inc\".",
      "required": true,
      "type": "string"
    },
    "applicationId": {
      "name": "applicationId",
      "in": "path",
      "description": "An app id, e.g. \"ReferenceBot\".",
      "required": true,
      "type": "string"
    },
...
    "accountId": {
      "name": "accountId",
      "in": "path",
      "description": "An account id, e.g. \"ReferenceBot\".",
      "required": true,
      "type": "string"
    },
    "upfile": {
      "in": "formData",
      "name": "upfile",
      "type": "file",
      "required": true,
      "description": "The file to upload."
    },
...
    "projectNameInForm": {
      "in": "formData",
      "name": "projectName",
      "type": "string",
      "minLength": 4,
      "maxLength": 50,
      "required": true,
      "description": "The name of the project"
    },
    "datasetNameInForm": {
      "in": "formData",
      "name": "datasetName",
      "type": "string",
      "minLength": 5,
      "maxLength": 50,
      "required": true,
      "description": "The name of the dataset"
    },
...
    "projectDescriptionInForm": {
      "in": "formData",
      "name": "projectDescription",
      "type": "string",
      "minLength": 1,
      "maxLength": 150,
      "required": false,
      "description": "The project description"
    },
...
    "projectId": {
      "name": "projectId",
      "in": "path",
      "description": "id of the project",
      "required": true,
      "type": "string"
    },
    "projectStatusEnabled": {
      "name": "projectStatusEnabled",
      "in": "query",
      "description": "project status ENABLED or DISABLED",
      "required": false,
      "type": "boolean",
      "default": true
    },
...
  },
  "responses": {
    "BadRequest": {
      "description": "The request was invalid. For example, the value of the initial parameter was invalid",
      "schema": {
        "type": "string"
      }
    },
    "NotFound": {
      "description": "The URI requested is invalid or the resource requested, such as a user, does not exist. Do not retry with the same input",
      "schema": {
        "type": "string"
      }
    },
    "Unauthorized": {
      "description": "Missing or incorrect authentication credentials. This may also returned in other undefined circumstances.Do not retry with the same input",
      "schema": {
        "type": "string"
      }
    },
    "Forbidden": {
      "description": "Forbidden. The request is understood, but it has been refused or access is not allowed. An accompanying error message will explain further. Do not retry with the same input",
      "schema": {
        "type": "string"
      }
    },
...
    "InternalServerError": {
      "description": "Internal server error",
      "schema": {
        "type": "string"
      }
    },
    "Conflict": {
      "description": "The request could not be completed due to a conflict with the current state of the resource.Do not retry with the same input.",
      "schema": {
        "type": "string"
      }
    }
  },
  "paths": {
...
    "/clients/{clientId}/accounts/{accountId}/applications/{applicationId}/projects": {
      "post": {
        "summary": "Creates a new project for a given client, app and account",
        "x-swagger-router-controller": "projects",
        "operationId": "createProject",
        "tags": [
          "project"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/upfile"
          },
          {
            "$ref": "#/parameters/clientId"
          },
          {
            "$ref": "#/parameters/applicationId"
          },
          {
            "$ref": "#/parameters/accountId"
          },
          {
            "$ref": "#/parameters/projectNameInForm"
          },
          {
            "$ref": "#/parameters/datasetNameInForm"
          },
          {
            "$ref": "#/parameters/projectDescriptionInForm"
          }
        ],
        "responses": {
          "200": {
            "description": "A new project was created successfully.",
            "schema": {
              "$ref": "#/definitions/ProjectAfterCreation"
            }
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          },
          "400": {
            "$ref": "#/responses/BadRequest"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          },
          "403": {
            "$ref": "#/responses/Forbidden"
          }
        }
      },
      "get": {
        "summary": "Get all the projects for a given app",
        "x-swagger-router-controller": "projects",
        "operationId": "getProjects",
        "tags": [
          "project"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/clientId"
          },
          {
            "$ref": "#/parameters/applicationId"
          },
          {
            "$ref": "#/parameters/accountId"
          },
          {
            "$ref": "#/parameters/projectStatusEnabled"
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieved Projects",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Project"
              }
            }
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/responses/Forbidden"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      }
    },
    "/clients/{clientId}/accounts/{accountId}/applications/{applicationId}/projects/{projectId}": {
      "patch": {
        "summary": "Updates the project name and/or project description for a given client, app, account and projectId",
        "x-swagger-router-controller": "projects",
        "operationId": "updateProject",
        "tags": [
          "project"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/clientId"
          },
          {
            "$ref": "#/parameters/applicationId"
          },
          {
            "$ref": "#/parameters/accountId"
          },
          {
            "$ref": "#/parameters/projectId"
          },
          {
            "in": "body",
            "name": "Project name",
            "description": "The project to update",
            "schema": {
              "type": "object",
              "properties": {
                "projectName": {
                  "type": "string",
                  "minLength": 4,
                  "maxLength": 50
                },
                "projectDescription": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 150
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated a project",
            "schema": {
              "$ref": "#/definitions/Project"
            }
          },
          "400": {
            "$ref": "#/responses/BadRequest"
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/responses/Forbidden"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          },
          "409": {
            "$ref": "#/responses/Conflict"
          }
        }
      },
      "delete": {
        "summary": "Disables the project for a given client, app, account and projectId",
        "x-swagger-router-controller": "projects",
        "operationId": "deleteProject",
        "tags": [
          "project"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/clientId"
          },
          {
            "$ref": "#/parameters/applicationId"
          },
          {
            "$ref": "#/parameters/accountId"
          },
          {
            "$ref": "#/parameters/projectId"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted a project",
            "schema": {
              "$ref": "#/definitions/Project"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/responses/Forbidden"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          }
        }
      },
      "get": {
        "summary": "Get the project for a given app",
        "x-swagger-router-controller": "projects",
        "operationId": "getProject",
        "tags": [
          "project"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/clientId"
          },
          {
            "$ref": "#/parameters/applicationId"
          },
          {
            "$ref": "#/parameters/accountId"
          },
          {
            "$ref": "#/parameters/projectId"
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieved Project",
            "schema": {
              "$ref": "#/definitions/Project"
            }
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/responses/Forbidden"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ProjectAfterCreation": {
      "description": "Project Object just after creating a project",
      "type": "object",
      "properties": {
        "applicationId": {
          "description": "The app id of the Project",
          "type": "string"
        },
        "clientId": {
          "description": "The client id of the Project",
          "type": "string"
        },
        "accountId": {
          "description": "The account id of the Project",
          "type": "string"
        },
        "projectName": {
          "description": "The name of the project",
          "type": "string"
        },
        "datasetName": {
          "description": "The name of the project",
          "type": "string"
        },
        "projectDescription": {
          "description": "The description of the project",
          "type": "string"
        },
        "createdBy": {
          "description": "The id of the person who created the project",
          "type": "string"
        },
        "status": {
          "description": "The status of the project",
          "type": "string"
        }
      }
    },
...
    "Project": {
      "description": "Project Object",
      "type": "object",
      "properties": {
        "applicationId": {
          "description": "The app id of the Project",
          "type": "string"
        },
        "clientId": {
          "description": "The client id of the Project",
          "type": "string"
        },
        "accountId": {
          "description": "The account id of the Project",
          "type": "string"
        },
        "projectName": {
          "description": "The name of the project",
          "type": "string"
        },
        "datasetName": {
          "description": "The name of the project",
          "type": "string"
        },
        "projectDescription": {
          "description": "The description of the project",
          "type": "string"
        },
        "modified": {
          "description": "The UTC timestamp when the project was modified",
          "type": "number"
        },
        "modifiedBy": {
          "description": "The id of the person who modified the project",
          "type": "string"
        },
        "created": {
          "description": "The UTC timestamp when the project was created",
          "type": "number"
        },
        "createdBy": {
          "description": "The id of the person who created the project",
          "type": "string"
        },
        "projectStatus": {
          "description": "The status of the project",
          "type": "string"
        },
        "projectStatusDescription": {
          "description": "The description of the status of the project",
          "type": "string"
        }
      }
    },
...
  }
}